fix: change to typescript
