+++ src/index.ts
@@ -3,6 +3,7 @@ import express, { Request, Response } from "express"
 import handlebars from "express-handlebars"
 import * as fs from "fs"
 import * as https from "https"
+import "./i18n"
 
 import { middleware as middlewareLogger } from "./pkg/logger"
 import { toUiNodePartial } from "./pkg/ui"
